fix(fantams): les constantes et variables portent des reels
Cinquieme des sept causes de divergence. 105 -> 107 sources identiques a rasm au
bit pres, 8 -> 6 divergences, zero regression, 226 tests verts. synchro1 et
synchro2-2lal, qui divergeaient sur une table de 256 octets entiere, sont
desormais exactes.

L'evaluateur calculait deja en double, mais le type se perdait au passage par un
NOM : expr::Resolver rendait un int64_t et expr::Result ne portait qu'une valeur
arrondie. « v = 2.4 » puis « db v*2 » donnait donc 4 la ou rasm donne 5 —
l'ecrasement ayant lieu au stockage, pas au calcul. L'erreur n'avait meme pas de
biais constant : 2.4 partait vers le bas, 2.5 vers le haut.

  v = 2.4 / w = 2.5 / x = 2.6, puis db v*2, w*2, x*2
    avant : 04 06 06        apres : 05 05 05        rasm : 05 05 05

expr::Result porte desormais `real` a cote de `value`, et le resolveur rend un
double. La table des symboles de l'assembleur est en double ; asmb::Output
l'arrondit a la sortie, car c'est une table d'ADRESSES destinee aux outils — la
precision reelle n'a d'interet qu'a l'interieur du calcul. C'est l'ADR 0008
applique jusqu'au bout, la ou il ne l'etait qu'a moitie.

Piege evite : la reinjection d'une valeur dans du TEXTE source passe par fmtNum,
qui rend la forme entiere quand la valeur l'est. Sans ca, « lab{i} » dans une
boucle aurait produit « lab1.000000 ».

Reste sur les 6 divergences : 3 relevent du wrapper hasLiteOrg (hote, pas
l'assembleur), 2 de l'ecart d'un ULP entre le double de fantams et le long
double de rasm — divergence assumee — et 1 non instruite (Dots).

+++ fantams/expr_test.cpp
@@ -7,8 +7,8 @@
 
 static int g_pass = 0, g_fail = 0;
 
-static std::map<std::string, int64_t> g_syms;
-static expr::Resolver resolver = [](const std::string &n, int64_t &out) -> bool {
+static std::map<std::string, double> g_syms;
+static expr::Resolver resolver = [](const std::string &n, double &out) -> bool {
     auto it = g_syms.find(n);
     if (it == g_syms.end()) return false;
     out = it->second; return true;
